Add permission check in removeById method of ProjectService

+++ tasks-app/src/main/java/com/tasks/business/ProjectService.java
@@ -4,6 +4,8 @@ import com.tasks.business.entities.Project;
 import com.tasks.business.entities.User;
 import com.tasks.business.exceptions.DuplicatedResourceException;
 import com.tasks.business.exceptions.InstanceNotFoundException;
+import com.tasks.business.exceptions.PermissionException;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.tasks.business.repository.ProjectsRepository;
@@ -63,10 +65,13 @@ public class ProjectService {
     }
     
     @Transactional()
-    public void removeById(Long id) throws InstanceNotFoundException {
+    public void removeById(String userName , Long id) throws InstanceNotFoundException {
         Optional<Project> project = projectsRepository.findById(id);
         if(!project.isPresent()) {
             throw new InstanceNotFoundException(id, "Project" , MessageFormat.format("Project {0} does not exist", id));
+        }
+        if(!project.get().getAdmin().getUsername().equals(userName)) {
+            throw new PermissionException();
         }
         projectsRepository.delete(project.get());
     }
